Planner: extract resolver module
